Created datasources package and renamed utils
