+++ App/src/main/AndroidManifest.xml
@@ -56,7 +56,7 @@
         <service android:name=".services.DataService" />
 
         <receiver
-            android:name=".utils.gcm.GcmBroadcastReceiver"
+            android:name=".utilities.gcm.GcmBroadcastReceiver"
             android:permission="com.google.android.c2dm.permission.SEND">
             <intent-filter>
                 <action android:name="com.google.android.c2dm.intent.RECEIVE" />
