+++ App/src/main/java/cc/softwarefactory/lokki/android/fragments/ContactsFragment.java
@@ -23,8 +23,8 @@ import com.androidquery.AQuery;
 import cc.softwarefactory.lokki.android.MainApplication;
 import cc.softwarefactory.lokki.android.R;
 import cc.softwarefactory.lokki.android.avatar.AvatarLoader;
-import cc.softwarefactory.lokki.android.utils.PreferenceUtils;
-import cc.softwarefactory.lokki.android.utils.Utils;
+import cc.softwarefactory.lokki.android.utilities.PreferenceUtils;
+import cc.softwarefactory.lokki.android.utilities.Utils;
 
 import org.json.JSONArray;
 import org.json.JSONException;
