+++ App/src/main/java/cc/softwarefactory/lokki/android/fragments/PlacesFragment.java
@@ -25,12 +25,12 @@ import android.widget.ListView;
 
 import com.androidquery.AQuery;
 
-import cc.softwarefactory.lokki.android.utils.ServerApi;
+import cc.softwarefactory.lokki.android.utilities.ServerApi;
 import cc.softwarefactory.lokki.android.services.DataService;
 import cc.softwarefactory.lokki.android.MainApplication;
 import cc.softwarefactory.lokki.android.R;
 import cc.softwarefactory.lokki.android.avatar.AvatarLoader;
-import cc.softwarefactory.lokki.android.utils.PreferenceUtils;
+import cc.softwarefactory.lokki.android.utilities.PreferenceUtils;
 
 import com.makeramen.RoundedImageView;
