+++ App/src/main/java/cc/softwarefactory/lokki/android/fragments/SettingsFragment.java
@@ -17,13 +17,13 @@ import android.widget.ImageView;
 
 import com.androidquery.AQuery;
 
-import cc.softwarefactory.lokki.android.utils.ServerApi;
+import cc.softwarefactory.lokki.android.utilities.ServerApi;
 import cc.softwarefactory.lokki.android.services.LocationService;
 import cc.softwarefactory.lokki.android.MainApplication;
 import cc.softwarefactory.lokki.android.R;
 import cc.softwarefactory.lokki.android.avatar.AvatarLoader;
-import cc.softwarefactory.lokki.android.utils.PreferenceUtils;
-import cc.softwarefactory.lokki.android.utils.Utils;
+import cc.softwarefactory.lokki.android.utilities.PreferenceUtils;
+import cc.softwarefactory.lokki.android.utilities.Utils;
 
 
 public class SettingsFragment extends Fragment implements AdapterView.OnItemSelectedListener {
