+++ App/src/main/java/cc/softwarefactory/lokki/android/utilities/map/MapUserTypes.java
@@ -1,4 +1,4 @@
-package cc.softwarefactory.lokki.android.utils.map;
+package cc.softwarefactory.lokki.android.utilities.map;
 
 /**
  * Created by verne_000 on 16.2.2015.
