+++ App/src/main/java/cc/softwarefactory/lokki/android/utilities/map/MapUtils.java
@@ -1,4 +1,4 @@
-package cc.softwarefactory.lokki.android.utils.map;
+package cc.softwarefactory.lokki.android.utilities.map;
 
 import android.location.Location;
