+++ App/src/test/java/cc/softwarefactory/lokki/android/espresso/AboutScreenTest.java
@@ -9,7 +9,7 @@ import android.os.PatternMatcher;
 
 import cc.softwarefactory.lokki.android.R;
 import cc.softwarefactory.lokki.android.espresso.utilities.TestUtils;
-import cc.softwarefactory.lokki.android.utils.Utils;
+import cc.softwarefactory.lokki.android.utilities.Utils;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
