+++ App/src/test/java/cc/softwarefactory/lokki/android/espresso/utilities/TestUtils.java
@@ -12,7 +12,7 @@ import android.view.View;
 
 import cc.softwarefactory.lokki.android.activities.MainActivity;
 import cc.softwarefactory.lokki.android.R;
-import cc.softwarefactory.lokki.android.utils.PreferenceUtils;
+import cc.softwarefactory.lokki.android.utilities.PreferenceUtils;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
